+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
- - ✨ Added the ability to delete AOIs with the `delete` key
+ - ✨ Added the ability to delete AOIs with the `delete` key.
+ - ✨ Added detector and track projected lat / lon caching to drastically speed up viewing these overlays in the geographic view.
  - 🛠️ Updated track plot window to reset zoom view when axes are changed.
  - 🛠️ Updated projection cache default.
  - 🛠️ Updated data loader to allow loading a slice of imagery incrementally.
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - 🔧 Fixed edge-case bug where small images might not be able to project into the geographic view.
  - 🔧 Corrected where preventing histogram plot updates takes effect when histogram is hidden.
+ - 🔧 Fixed bug where track points from all frames were projected using the current frame's geolocation metadata
 
 ## [1.10.0] - 2026-2-15